requirements.txt
asgiref==3.11.0
Django==4.2.27
djangorestframework==3.16.1
dotenv==0.9.9
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
